Creat R code (P8–P14)
This part of the script is based on the tutorial (pages P8–P14). Most of the code follows the tutorial structure, with additional comments  added to improve readability and understanding.
The function design.explore was modified to better reflect our use case, including the use of real intervention durations and the distinction between major and minor maintenance events. Further work is required to complete the model, including:
- adding the missing input data for all systems, and
- extending the current functions so that they can integrate six systems  instead of only two.